wtc's second patch for 158704 1. Link the .so's with -lc only on Solaris.  (SunOS 4.1.3 doesn't need this change.)
2. Add libVersionPoint to NSPR_4.0 in the mapfiles.

3. Try to do ifdef USE_64 in mozilla/nsprpub/pkg/solaris/Makefile
in the way Danek suggested.  I had to use :: (as opposed to :) to
specify PROTO's dependency because :: dependency can be used
multiple times for the same target.

4. Fixed a bug in mozilla/nsprpub/pkg/solaris/Makefile in
the creation of the symlink $(ROOT)/usr/lib/mps/sparcv9.

5. Removed the symbols that shouldn't be exported from nspr.def.
Added PR_StackPop and PR_StackPush (which the 'nm' command
reports as 'N', not 'T').  Added the PRP_* functions.

--HG--
branch : NSPRPUB_RELEASE_4_1_2_SUN_PKG_BRANCH